requirements.txt
anthropic
langgraph
langchain
langchain-anthropic
pandas
streamlit
plotly
rapidfuzz
sentence-transformers
chromadb
python-dotenv
faker
fpdf2
numpy
scikit-learn